First rendering of Media Repository Module
This is the output of Module Studio. I did clean up the interface for
the templates. I know the correct way to do this is to have these in
the themes templates and then have them override what is in the module.
For ease of updating. I am not doing that right now. I will fix that
later.